+++ common/src/main/resources/assets/zombie_tactics/icon.txt
@@ -1,0 +1,2 @@
+Icon image is from
+https://minecraft.wiki/w/Zombie#/media/File:Zombie_JE3_BE2.png
