+++ fabric/src/main/java/n643064/zombie_tactics/fabric/client/MainFabricClient.java
@@ -1,0 +1,10 @@
+package n643064.zombie_tactics.fabric.client;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public final class MainFabricClient implements ClientModInitializer {
+    @Override
+    public void onInitializeClient() {
+        // This entrypoint is suitable for setting up client-specific logic, such as rendering.
+    }
+}
