+++ fabric/src/main/resources/fabric.mod.json
@@ -1,0 +1,35 @@
+{
+  "schemaVersion": 1,
+  "id": "zombie_tactics",
+  "version": "${version}",
+  "name": "ZombieTactics",
+  "description": "Improves zombie mechanism",
+  "authors": ["n643064"],
+  "contact": {
+    "homepage": ""
+  },
+  "license": "GPL-3.0",
+  "icon": "assets/zombie_tactics/icon.png",
+  "environment": "*",
+  "entrypoints": {
+    "main": [
+      "n643064.zombie_tactics.fabric.Main"
+    ],
+    "client": [
+      "n643064.zombie_tactics.fabric.client.MainFabricClient"
+    ]
+  },
+  "mixins": [
+    "zombie_tactics.mixins.json"
+  ],
+  "depends": {
+    "fabricloader": ">=0.16.10",
+    "minecraft": "~1.21.1",
+    "java": ">=21"
+    ,"architectury": ">=13.0.8"
+    ,"fabric-api": "*"
+  },
+  "suggests": {
+    "another-mod": "*"
+  }
+}
