+++ gradle.properties
@@ -1,40 +1,18 @@
-# Sets default memory used for gradle commands. Can be overridden by user or command line properties.
+# Done to increase the memory available to Gradle.
 org.gradle.jvmargs=-Xmx2G
-org.gradle.daemon=true
 org.gradle.parallel=true
-org.gradle.caching=true
-org.gradle.configuration-cache=true
-## Environment Properties
-# You can find the latest versions here: https://projects.neoforged.net/neoforged/neoforge
-# The Minecraft version must agree with the Neo version to get a valid artifact
-minecraft_version=1.21.1
-# The Minecraft version range can use any release version of Minecraft as bounds.
-# Snapshots, pre-releases, and release candidates are not guaranteed to sort properly
-# as they do not follow standard versioning conventions.
-minecraft_version_range=[1.21,1.22)
-# The Neo version must agree with the Minecraft version to get a valid artifact
-neo_version=21.1.138
-# The Neo version range can use any version of Neo as bounds
-neo_version_range=[21,)
-# The loader version range can only use the major version of FML as bounds
-loader_version_range=[4,)
-parchment_minecraft_version=1.21.1
-parchment_mappings_version=2024.11.17
-## Mod Properties
-# The unique mod identifier for the mod. Must be lowercase in English locale. Must fit the regex [a-z][a-z0-9_]{1,63}
-# Must match the String constant located in the main mod class annotated with @Mod.
-mod_id=zombie_tactics
-# The human-readable display name for the mod.
-mod_name=Zombie Tactics
-# The license of the mod. Review your options at https://choosealicense.com/. All Rights Reserved is the default.
-mod_license=GNU GPL 3.0
-# The mod version. See https://semver.org/
-mod_version=1.3.0
-# The group ID for the mod. It is only important when publishing as an artifact to a Maven repository.
-# This should match the base package used for the mod sources.
-# See https://maven.apache.org/guides/mini/guide-naming-conventions.html
-mod_group_id=n643064
-# The authors of the mod. This is a simple text string that is used for display purposes in the mod list.
-mod_authors=n643064
-# The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
-mod_description=Improves zombie AI
+
+# Mod properties
+mod_version = 1.3.0
+maven_group = n643064
+archives_name = zombie_tactics
+enabled_platforms = fabric,neoforge
+
+# Minecraft properties
+minecraft_version = 1.21.1
+
+# Dependencies
+architectury_api_version = 13.0.8
+fabric_loader_version = 0.16.10
+fabric_api_version = 0.115.3+1.21.1
+neoforge_version = 21.1.140
