+++ neoforge/src/main/resources/META-INF/neoforge.mods.toml
@@ -1,0 +1,38 @@
+modLoader = "javafml"
+loaderVersion = "[2,)"
+#issueTrackerURL = ""
+license = "GNU GPL 3.0"
+
+[[mods]]
+modId = "zombie_tactics"
+version = "${version}"
+displayName = "ZombieTactics"
+authors = "n643064"
+description = '''
+Improves zombie mechanism
+'''
+#logoFile = ""
+
+[[dependencies.zombie_tactics]]
+modId = "neoforge"
+type = "required"
+versionRange = "[21,)"
+ordering = "NONE"
+side = "BOTH"
+
+[[dependencies.zombie_tactics]]
+modId = "minecraft"
+type = "required"
+versionRange = "[1.21.1,)"
+ordering = "NONE"
+side = "BOTH"
+
+[[dependencies.zombie_tactics]]
+modId = "architectury"
+type = "required"
+versionRange = "[13.0.8,)"
+ordering = "AFTER"
+side = "BOTH"
+
+[[mixins]]
+config = "zombie_tactics.mixins.json"
