+++ neoforge/src/main/resources/zombie_tactics.mixins.json
@@ -1,0 +1,16 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "n643064.zombie_tactics.neoforge.mixin",
+  "compatibilityLevel": "JAVA_21",
+  "refmap": "zombie_tactics.refmap.json",
+  "mixins": [
+    "ZombieAttackGoalMixin",
+    "ZombieMixin"
+  ],
+  "client": [
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
