+++ settings.gradle
@@ -1,11 +1,14 @@
 pluginManagement {
     repositories {
-        mavenLocal()
+        maven { url "https://maven.fabricmc.net/" }
+        maven { url "https://maven.architectury.dev/" }
+        maven { url "https://files.minecraftforge.net/maven/" }
         gradlePluginPortal()
-        maven { url = 'https://maven.neoforged.net/releases' }
     }
 }
 
-plugins {
-    id 'org.gradle.toolchains.foojay-resolver-convention' version '0.8.0'
-}
+rootProject.name = 'zombie_tactics'
+
+include 'common'
+include 'fabric'
+include 'neoforge'
