--- src/main/resources/zombie_tactics.mixins.json
@@ -1,17 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "n643064.zombie_tactics.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "refmap": "zombie_tactics.refmap.json",
-  "mixins": [
-    "MeleeAttackGoalMixin",
-    "ZombieAttackGoalMixin",
-    "ZombieMixin"
-  ],
-  "client": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
